requirements.txt
flask>=2.0.1
google-cloud-bigquery>=3.11.4
pandas>=1.4.0
openpyxl>=3.0.9
google-auth>=2.12.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.12.0
python-dotenv>=0.19.0
pytest>=7.0.0
